Extract context to separate file
