+++ packages/pipe/src/createElement.ts
@@ -1,9 +1,5 @@
-import { BehaviorSubject, filter, Observable, Subject, switchMap, takeUntil } from 'rxjs';
-
-export type Context = {
-    get: <T>(key: string) => Observable<T>;
-    set: <T>(key: string, value: T) => void;
-};
+import { Observable, Subject, takeUntil } from 'rxjs';
+import { Context, initializeContext } from './Context';
 
 export type Component<Props extends Record<string, Observable<unknown> | unknown>> = (
     props: Props,
@@ -33,35 +29,8 @@ export function createElement<
     props: Props,
     children?: PipeNode[] | Observable<[string, PipeNode | null]>,
 ): PipeNode {
-    const contextValues$ = new Subject<[string, unknown]>();
-
-    const mappedObservables = new Map<string, Subject<unknown>>();
-
     const cleanup$ = new Subject<void>();
-    const contextListener$ = new BehaviorSubject(mappedObservables);
-
-    contextValues$.subscribe(([nextKey, nextValue]) => {
-        if (!mappedObservables.has(nextKey)) {
-            mappedObservables.set(nextKey, new BehaviorSubject(nextValue));
-        } else {
-            mappedObservables.get(nextKey)!.next(nextValue);
-        }
-        contextListener$.next(mappedObservables);
-    });
-
-    const context: Context = {
-        get: <T>(key: string) => {
-            return contextListener$.pipe(
-                filter((mappedObservables) => mappedObservables.has(key)),
-                switchMap((mappedObservables) => {
-                    return mappedObservables.get(key) as Observable<T>;
-                }),
-            );
-        },
-        set: <T>(key: string, value: T) => {
-            contextValues$.next([key, value]);
-        },
-    };
+    const { context, contextValues$ } = initializeContext(cleanup$);
 
     const element =
         typeof component === 'string'
@@ -71,10 +40,6 @@ export function createElement<
     cleanup$.subscribe({
         complete: () => {
             element.remove();
-            contextValues$.complete();
-            for (const observable of mappedObservables.values()) {
-                observable.complete();
-            }
         },
     });
